Less crashes. Editor registered correctly. Minor refactoring. And partial editor GUI
The Editor was crashing because it was still function stubs, so it returned a nullptr when requesting the path of the edited resource.
Editor suddenly registering just fine. IDK if it was a linking issue. Rebuilding entire project solved it?
Refactored to follow Godot's guidelines of snake_case filenames and avoiding pragma once in header files.
Starting work on getting the editor graphics to have some basic functionality. Inspired by Blender's UI to have a toggleable pane on the side with tabs for contextual information.