Refactored c11r_lang.*; Added an icon?; Got editor visuals working partially; Started work on block registry
Block registry is going to be a big deal for this language. The different ways to add functionality to an existing script will be really helpful for quickly developing functional scripts. I'm planning on focusing on a Design By Composition approach, which is pretty central to Godot as a whole anyway. So far I have 3 planned methods of adding custom functionality besides the internal graph.
Ideally, these compositions would allow getting an FPS controller up and running almost as a one-liner. And these compositing scripts could be made and shared elsewhere for beginners to use should they choose.