Refactored back to uberfile c11r_lang; Added BlockEntry; Working on adding blocks to graph
I'm not fully sure why extracting the sections into separate files broke it. That's probably a product of having more experience in Java/C# where separating into individual files is almost a requirement. C++ doesn't seem to care all that much. I'll just have to make better use of code editor tools to navigate these larger files. VSCode's outliner and sticky-scroll are life-savers for this.
Need to create a graph updating system. VisualScript guts the graph every time a new node is added. Which works, and shouldn't be a problem on smaller graphs. I'll have to figure out if the VisualScript option is best, or if I can come up with a better way to do it. So far, VisualScript seems to have a pretty solid structure for these back-end things. And I'm probably going to keep carrying over elements as I get this prototype working. Then I can start scrapping features that aren't helpful and formulate a better process. Big thing I wanna revise is execution order. That can and should be evaluated dynamically.