RIP AND TEAR
Clearing out a lot of the copypasta code out from the language. It was helpful to get stuff up and running quickly, but at this point it's causing more problems than its worth. I need to build from the ground up if I want to make something structurally different from VisualScript, which is definitely the goal <3