Getting the editor working better
Started out with fixing up the shortcuts so I can easily perform actions when the shortcuts are triggered. Then I started added an options menu with a few key functionalities needed from the editor. Currently working on filling out the "Change Base Script" option which will let the user change the base script easily. Ideally, I want this to also be set when creating a new script, but IDK how I want that to work as of yet